refactor(AgeDiscountPolicy): 클래스 내부에 상수를 관리하는 정적 클래스 추가 - fare -> path 패키지 이동 - 6세 미만인 경우의 요금을 계산하는 테스트케이스 추가하기
